Move back into original files
