+++ crates/aranya-util/Cargo.toml
@@ -14,7 +14,6 @@ workspace = true
 
 
 [dependencies]
-aranya-crypto = { workspace = true }
 aranya-fast-channels = { workspace = true }
 buggy = { workspace = true }
